Move daemon command to the root
