+++ cmd/root.go
@@ -7,6 +7,6 @@ import (
 // RootCmd represents the base command when called without any subcommands
 var RootCmd = &cobra.Command{
 	Use:               "mesg-core",
-	Short:             "MESG CORE",
+	Short:             "MESG Core",
 	DisableAutoGenTag: true,
 }
